First prototype using gopkg.in/yaml.v3, doesn't work well
Hmm, so the yaml.v3 is the problem here. Even when I force set the style
of the scalar string value node (with node.Style = yaml.LiteralStyle),
the yaml.Marshal() then doesn't respect it and sometimes it prints the
string as quoted string with escaped newlines \n instead..

This looks like a bug/limitation of the gopkg.in/yaml.v3 pkg. We'll need
to try some other approach.